backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv

# RAG / LLM
langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain_mistralai
groq
pypdf
google-genai
langchain-google-genai
pinecone
pyreadline3
# Auth & DB
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
email-validator